Refactor tests to a separate test/-dir.
