--- watch-check-meta-circularity.sh
@@ -1,2 +1,0 @@
-mkdir -p plantUML
-deno test --watch --allow-write src/m3/check-meta-circularity.ts
