+++ watch-test-m3.sh
@@ -1,0 +1,2 @@
+mkdir -p plantUML
+deno test --watch --allow-write src/m3/test/test-*.ts
